sponza_backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
bcrypt==4.1.3
billiard==4.2.1
blinker==1.8.2
cachelib==0.9.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.6.1
eventlet==0.38.0
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
graphviz==0.20.3
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==2.1.5
pillow==10.4.0
prompt_toolkit==3.0.48
pydot==3.0.1
PyJWT==2.10.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.2.0
six==1.16.0
SQLAlchemy==2.0.31
sqlalchemy_schemadisplay==2.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
